Move max tests to dedicated page
